debug pokemon and teams routes, add six pokemon limit to teams

+++ backend/api/pokemon.js
@@ -7,7 +7,7 @@ const { authenticate } = require("./auth");
 //Get all pokemon (probably unecessary but good to have)
 router.get("/", async (req, res, next) => {
   try {
-    const {pokemon} = await prisma.pokemon.findMany();
+    const pokemon = await prisma.pokemon.findMany();
     res.json(pokemon);
   } catch (e) {
     next(e);
@@ -16,6 +16,13 @@ router.get("/", async (req, res, next) => {
 // Create a new pokemon
 router.post("/", authenticate, async (req, res, next) => {
   const { name, ability, moves, teamId } = req.body;
+  const team = await prisma.team.findUnique({ 
+    where: {id: +teamId},
+    include: { pokemon: true }
+  })
+  if (team.pokemon.length >=6){
+    next({ status: 400, message: "This team is full" })
+  }
   if (!name) {
     next({ status: 400, message: "Name must be provided" })
   }
@@ -24,6 +31,9 @@ router.post("/", authenticate, async (req, res, next) => {
   }
   if (!teamId) {
     next({ status: 400, message: "A Pokemon must be on a team" })
+  }
+  if (team.ownerId !== req.user.id) {
+    next({ status: 403, message: "This is not your team" })
   }
   try {
     const pokemon = await prisma.pokemon.create({
@@ -47,11 +57,14 @@ router.get("/:id", async (req, res, next) => {
       next({ status: 404, message: `pokemon with id ${id} not found` })
     }
 
-    //returns an array with the pokemon and a boolean value of if the user who is viewing it ons it
+    //returns an object with the pokemon and a boolean value of if the user who is viewing it ons it
     const team = await prisma.team.findUnique( { where: { id : pokemon.teamId } });
     const isOwner = (req.user.id === team.ownerId)
     //frontend can handle rendering options
-    res.json(pokemon, isOwner);
+    res.json({ 
+      pokemon,
+       isOwner 
+      } );
   } catch (e) {
     next(e);
   }
@@ -88,12 +101,18 @@ router.get("/:id", async (req, res, next) => {
 
 router.delete("/:id", authenticate, async (req, res, next) => {
   const { id } = req.params;
+  const { userId } = req.user.id
 
   try {
     // Check if the Pokemon exists
     const pokemon = await prisma.pokemon.findUnique({ where: { id: +id } });
     if (!pokemon) {
       return next({ status: 404,message: `Pokemon with id ${id} does not exist.` });
+    }
+    const team = await prisma.team.findUnique({ where: {id: +pokemon.teamId}})
+    
+    if (req.user.id !== team.ownerId) {
+      next({ status: 403, message: "This is not your team" })
     }
     await prisma.pokemon.delete({ where: { id: +id } });
     res.sendStatus(204);
